requirements.txt
flask
flask-cors
flask-limiter

#dotenv
zappa


boto3
logmatic-python
python-dotenv
requests

coverage
black
pyflakes
